--- a/aten/src/ATen/native/AffineGridGenerator.cpp
+++ b/aten/src/ATen/native/AffineGridGenerator.cpp
@@ -3,12 +3,16 @@
 
 namespace at { namespace native {
 
-at::Tensor linspace_from_neg_one(const Tensor& grid, int64_t num_steps) {
-  if (num_steps > 1) {
-    return at::linspace(-1, 1, num_steps, grid.options());
-  } else {
-    return at::tensor(-1, grid.options());
+at::Tensor linspace_from_neg_one(const Tensor& grid, int64_t num_steps, 
+                                 bool align_corners) {
+  if (num_steps <= 1) {
+    return at::tensor(0, grid.options());
   }
+  auto range = at::linspace(-1, 1, num_steps, grid.options());
+  if (!align_corners) {
+    range = range * (num_steps - 1) / num_steps;
+  }
+  return range;
 }
 
 Tensor make_base_grid_4D(
@@ -16,11 +20,12 @@ Tensor make_base_grid_4D(
     int64_t N,
     int64_t C,
     int64_t H,
-    int64_t W) {
+    int64_t W,
+    bool align_corners) {
   auto base_grid = at::empty({N, H, W, 3}, theta.options());
 
-  base_grid.select(-1, 0).copy_(linspace_from_neg_one(theta, W));
-  base_grid.select(-1, 1).copy_(linspace_from_neg_one(theta, H).unsqueeze_(-1));
+  base_grid.select(-1, 0).copy_(linspace_from_neg_one(theta, W, align_corners));
+  base_grid.select(-1, 1).copy_(linspace_from_neg_one(theta, H, align_corners).unsqueeze_(-1));
   base_grid.select(-1, 2).fill_(1);
 
   return base_grid;
@@ -32,12 +37,13 @@ Tensor make_base_grid_5D(
     int64_t C,
     int64_t D,
     int64_t H,
-    int64_t W) {
+    int64_t W,
+    bool align_corners) {
   auto base_grid = at::empty({N, D, H, W, 4}, theta.options());
 
-  base_grid.select(-1, 0).copy_(linspace_from_neg_one(theta, W));
-  base_grid.select(-1, 1).copy_(linspace_from_neg_one(theta, H).unsqueeze_(-1));
-  base_grid.select(-1, 2).copy_(linspace_from_neg_one(theta, D).unsqueeze_(-1).unsqueeze_(-1));
+  base_grid.select(-1, 0).copy_(linspace_from_neg_one(theta, W, align_corners));
+  base_grid.select(-1, 1).copy_(linspace_from_neg_one(theta, H, align_corners).unsqueeze_(-1));
+  base_grid.select(-1, 2).copy_(linspace_from_neg_one(theta, D, align_corners).unsqueeze_(-1).unsqueeze_(-1));
   base_grid.select(-1, 3).fill_(1);
 
   return base_grid;
@@ -50,7 +56,7 @@ Tensor affine_grid_generator_4D(
     int64_t H,
     int64_t W,
     bool align_corners) {
-  Tensor base_grid = make_base_grid_4D(theta, N, C, H, W);
+  Tensor base_grid = make_base_grid_4D(theta, N, C, H, W, align_corners);
   auto grid = base_grid.view({N, H * W, 3}).bmm(theta.transpose(1, 2));
   return grid.view({N, H, W, 2});
 }
@@ -63,7 +69,7 @@ Tensor affine_grid_generator_5D(
     int64_t H,
     int64_t W,
     bool align_corners) {
-  Tensor base_grid = make_base_grid_5D(theta, N, C, D, H, W);
+  Tensor base_grid = make_base_grid_5D(theta, N, C, D, H, W, align_corners);
   auto grid = base_grid.view({N, D * H * W, 4}).bmm(theta.transpose(1, 2));
   return grid.view({N, D, H, W, 3});
 }
@@ -88,7 +94,7 @@ Tensor affine_grid_generator_4D_backward(
     int64_t H,
     int64_t W,
     bool align_corners) {
-  auto base_grid = make_base_grid_4D(grad_grid, N, C, H, W);
+  auto base_grid = make_base_grid_4D(grad_grid, N, C, H, W, align_corners);
   AT_ASSERT(grad_grid.sizes() == IntArrayRef({N, H, W, 2}));
   auto grad_theta = base_grid.view({N, H * W, 3})
                         .transpose(1, 2)
@@ -104,7 +110,7 @@ Tensor affine_grid_generator_5D_backward(
     int64_t H,
     int64_t W,
     bool align_corners) {
-  auto base_grid = make_base_grid_5D(grad_grid, N, C, D, H, W);
+  auto base_grid = make_base_grid_5D(grad_grid, N, C, D, H, W, align_corners);
   AT_ASSERT(grad_grid.sizes() == IntArrayRef({N, D, H, W, 3}));
   auto grad_theta = base_grid.view({N, D * H * W, 4})
                         .transpose(1, 2)