… aten/src/ATen/native/cuda/GridSampler.cu
… aten/src/ATen/native/cuda/GridSampler.cuh
… test/test_nn.py
… tools/autograd/derivatives.yaml