@bnehoran 2019-08-07
I just thought I’d add that while working on this pull request, I discovered a bug in how gradients are treated at the border of the image for `border` and `reflection` modes, and I opened an issue for it at https://github.com/pytorch/pytorch/issues/23925.
Though much the same code is affected, the underlying issue seemed somewhat tangential to the `align_corners` feature, and so it didn’t seem quite right to combine it with this pull request, but I still thought I’d mention it here in case people have a different opinion.
For now, I’ve left this bug in place, as is, so that but for the functionality explicitly addressed here, the gradient calculation is the same as before.