@soumith 2019-08-07
@ssnl @fmassa please review BC-Breaking concerns.
@ailzhang please review code