@bnehoran 2019-08-10
The tests in `test_nn.py` have now been updated to test both the old `align_corners=True` and the  new `align_corners=False` versions of `grid_sample` and `affine_grid`, and they seem to be passing.

However, it still looks like a couple checks are failing: one with
```
 Traceback (most recent call last):
   File "<string>", line 1, in <module>
   File "/opt/python/cp27-cp27mu/lib/python2.7/site-packages/torch/__init__.py", line 81, in <module>
     from torch._C import *
 ImportError: libmklml_intel.so: cannot open shared object file: No such file or directory
```
and another with 
```
 Traceback (most recent call last):
   File "/var/lib/jenkins/workspace/xla/scripts/gen.py", line 1060, in <module>
     generate(args)
   File "/var/lib/jenkins/workspace/xla/scripts/gen.py", line 1025, in generate
     assert check_overrides(overrides, overridden)
 AssertionError
 Failed to run '['/var/lib/jenkins/workspace/xla/scripts/generate_code.sh']'
```
but I'm really at a loss about what to do with these errors.
Any help with pinning them down would be appreciated.

Otherwise, it's essentially ready for review.