@bnehoran 2019-08-16
I've now factored out the `grid_sampler` helper functions from the 3D CPU and CUDA implementations into `GridSampler.h` and `GridSampler.cuh` respectively, including the coordinate conversion logic.
I had already refactored the 2D CPU implementation, but since it is vectorized, it's factored in a very different way and can't really make use of the same helper functions.