@SsnL 2019-08-16
@bnehoran Thanks for the patch! Sorry for the late reply. 

The amount of kernel code changes is huge. I trust you and @ailzhang on this. The BC warning is good as well. I do have some questions regarding to the performance. Do you think you could perform some benchmarking so we can see if the code slows down on the old setting (`align_corners=True`) and see if the new default is slower than the old default, and if so, how much? A table similar to https://github.com/pytorch/pytorch/pull/10980#issuecomment-420407362 would be awesome.

One expected slow down is the CUDA default setting, since the old one calls into cuDNN, which should have hyper optimized kernels. But that is fine.