@bnehoran 2019-08-17
@SsnL I ran timing tests using [this](https://gist.github.com/bnehoran/3c916d88ec97bf61f18a88e5cb6c7268) script, and here is the timing table:

```
-------|-------------------|-------------------|----------|----------|-------|----------------------|----------------------|-------------------
device |    input shape    |    grid shape     |  interp  | padding  | align | time (ms): this PR   | master from source   | 1.2.0 from binary
-------|-------------------|-------------------|----------|----------|-------|----------------------|----------------------|-------------------
 cuda  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | bilinear |  zeros   | True  | 0.4626655578613281   | 0.4640769958496094   | 0.49254417419433594
 cuda  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | bilinear |  zeros   | False | 0.45885562896728516  |                      | 
 cuda  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | bilinear |  border  | True  | 0.4248380661010742   | 0.45383691787719727  | 0.453951358795166
 cuda  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | bilinear |  border  | False | 0.4208683967590332   |                      | 
 cuda  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | bilinear |reflection| True  | 0.4295039176940918   | 0.4445815086364746   | 0.48131227493286133
 cuda  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | bilinear |reflection| False | 0.39810895919799805  |                      | 
 cuda  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | nearest  |  zeros   | True  | 0.17013072967529297  | 0.176239013671875    | 0.1787424087524414
 cuda  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | nearest  |  zeros   | False | 0.1703333854675293   |                      | 
 cuda  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | nearest  |  border  | True  | 0.16873598098754883  | 0.17609357833862305  | 0.17376422882080078
 cuda  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | nearest  |  border  | False | 0.16863107681274414  |                      | 
 cuda  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | nearest  |reflection| True  | 0.20630359649658203  | 0.2175593376159668   | 0.21936893463134766
 cuda  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | nearest  |reflection| False | 0.21077394485473633  |                      | 
 cuda  | [32, 3, 244, 244] | [32, 244, 244, 2] | bilinear |  zeros   | True  | 0.7029366493225098   | 0.730435848236084    | 0.7439589500427246
 cuda  | [32, 3, 244, 244] | [32, 244, 244, 2] | bilinear |  zeros   | False | 0.6895995140075684   |                      | 
 cuda  | [32, 3, 244, 244] | [32, 244, 244, 2] | bilinear |  border  | True  | 0.6884360313415527   | 0.7224059104919434   | 0.7246208190917969
 cuda  | [32, 3, 244, 244] | [32, 244, 244, 2] | bilinear |  border  | False | 0.6936454772949219   |                      | 
 cuda  | [32, 3, 244, 244] | [32, 244, 244, 2] | bilinear |reflection| True  | 0.6937479972839355   | 0.6976437568664551   | 0.6975317001342773
 cuda  | [32, 3, 244, 244] | [32, 244, 244, 2] | bilinear |reflection| False | 0.6895136833190918   |                      | 
 cuda  | [32, 3, 244, 244] | [32, 244, 244, 2] | nearest  |  zeros   | True  | 0.21797657012939453  | 0.21735906600952148  | 0.21885395050048828
 cuda  | [32, 3, 244, 244] | [32, 244, 244, 2] | nearest  |  zeros   | False | 0.21802186965942383  |                      | 
 cuda  | [32, 3, 244, 244] | [32, 244, 244, 2] | nearest  |  border  | True  | 0.21750926971435547  | 0.21637439727783203  | 0.21817445755004883
 cuda  | [32, 3, 244, 244] | [32, 244, 244, 2] | nearest  |  border  | False | 0.2175593376159668   |                      | 
 cuda  | [32, 3, 244, 244] | [32, 244, 244, 2] | nearest  |reflection| True  | 0.22589683532714844  | 0.23144960403442383  | 0.2327585220336914
 cuda  | [32, 3, 244, 244] | [32, 244, 244, 2] | nearest  |reflection| False | 0.22713899612426758  |                      | 
  cpu  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | bilinear |  zeros   | True  | 4.76588249206543     | 3.9052963256835938   | 5.741662979125977
  cpu  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | bilinear |  zeros   | False | 4.543666839599609    |                      | 
  cpu  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | bilinear |  border  | True  | 4.431045055389404    | 3.7845993041992188   | 4.976377487182617
  cpu  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | bilinear |  border  | False | 4.549868106842041    |                      | 
  cpu  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | bilinear |reflection| True  | 4.712920188903809    | 3.9714694023132324   | 5.6511759757995605
  cpu  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | bilinear |reflection| False | 5.3394389152526855   |                      | 
  cpu  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | nearest  |  zeros   | True  | 1.6000962257385254   | 1.0791325569152832   | 1.6270875930786133
  cpu  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | nearest  |  zeros   | False | 1.4397311210632324   |                      | 
  cpu  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | nearest  |  border  | True  | 1.4949965476989746   | 1.4754843711853027   | 1.7276549339294434
  cpu  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | nearest  |  border  | False | 1.565721035003662    |                      | 
  cpu  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | nearest  |reflection| True  | 2.05416202545166     | 1.8916535377502441   | 1.7394161224365234
  cpu  | [1024, 1, 28, 28] | [1024, 56, 56, 2] | nearest  |reflection| False | 2.2370004653930664   |                      | 
  cpu  | [32, 3, 244, 244] | [32, 244, 244, 2] | bilinear |  zeros   | True  | 8.025901317596436    | 7.095236778259277    | 8.783092498779297
  cpu  | [32, 3, 244, 244] | [32, 244, 244, 2] | bilinear |  zeros   | False | 7.528080940246582    |                      | 
  cpu  | [32, 3, 244, 244] | [32, 244, 244, 2] | bilinear |  border  | True  | 6.205465793609619    | 7.550709247589111    | 9.399261474609375
  cpu  | [32, 3, 244, 244] | [32, 244, 244, 2] | bilinear |  border  | False | 6.033613681793213    |                      | 
  cpu  | [32, 3, 244, 244] | [32, 244, 244, 2] | bilinear |reflection| True  | 6.127791404724121    | 6.305532455444336    | 8.782689571380615
  cpu  | [32, 3, 244, 244] | [32, 244, 244, 2] | bilinear |reflection| False | 6.14396333694458     |                      | 
  cpu  | [32, 3, 244, 244] | [32, 244, 244, 2] | nearest  |  zeros   | True  | 2.6335644721984863   | 2.660834789276123    | 4.572601318359375
  cpu  | [32, 3, 244, 244] | [32, 244, 244, 2] | nearest  |  zeros   | False | 2.6523566246032715   |                      | 
  cpu  | [32, 3, 244, 244] | [32, 244, 244, 2] | nearest  |  border  | True  | 2.2953200340270996   | 2.740752696990967    | 4.597971439361572
  cpu  | [32, 3, 244, 244] | [32, 244, 244, 2] | nearest  |  border  | False | 2.633633613586426    |                      | 
  cpu  | [32, 3, 244, 244] | [32, 244, 244, 2] | nearest  |reflection| True  | 2.4132323265075684   | 3.534998893737793    | 5.608177185058594
  cpu  | [32, 3, 244, 244] | [32, 244, 244, 2] | nearest  |reflection| False | 2.894620895385742    |                      | 
```